refactor(settings): extract dev_tools.dart
