+++ lib/me/settings/settings.dart
@@ -3,40 +3,16 @@ import "dart:io";
 import "package:flutter/material.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
 import "package:functional_widget_annotation/functional_widget_annotation.dart";
-import "package:logging/logging.dart";
 import "package:restart_app/restart_app.dart";
-import "package:riverpod_annotation/riverpod_annotation.dart";
 
 import "../../generated/l10n.dart";
-import "../../utils/database.dart";
-import "../../utils/debug/logger.dart";
-import "../../utils/dio.dart";
 import "../../utils/strings.dart";
-import "../../utils/urls.dart";
 import "chart_settings.dart";
+import "dev_tools.dart";
 import "providers.dart";
 import "section_title.dart";
 
 part "settings.g.dart";
-
-@riverpod
-class _LoggerLevel extends _$LoggerLevel {
-  @override
-  Level build() {
-    final index = prefs.getInt(K.loggerLevelIndex) ?? infoLevelIndex;
-    return loggerLevels[index];
-  }
-
-  Future<void> set(Level level) async {
-    state = level;
-    Logger.root.level = level;
-
-    final index = loggerLevels.indexOf(level);
-    await prefs.setInt(K.loggerLevelIndex, index);
-  }
-
-  Future<void> setIndex(int index) => set(loggerLevels[index]);
-}
 
 @cwidget
 Widget _settings(BuildContext context, WidgetRef ref) {
@@ -52,8 +28,6 @@ Widget _settings(BuildContext context, WidgetRef ref) {
 
   // devTools settings
   final showDevTools = ref.watch(showDevToolsProvider);
-  final loggerLevel = ref.watch(_loggerLevelProvider);
-  final loggerLevelIndex = loggerLevels.indexOf(loggerLevel);
 
   return Scaffold(
     appBar: AppBar(title: Text(s.settings)),
@@ -195,37 +169,7 @@ Widget _settings(BuildContext context, WidgetRef ref) {
             await ref.read(showDevToolsProvider.notifier).set(on);
           },
         ),
-        if (showDevTools)
-          SwitchListTile.adaptive(
-            secondary: const Icon(Icons.device_hub_outlined),
-            title: Text(s.fakeDevice),
-            value: ref.watch(fakeDeviceOnProvider),
-            onChanged: ref.read(fakeDeviceOnProvider.notifier).set,
-          ),
-        if (showDevTools)
-          ListTile(
-            leading: const Icon(Icons.text_snippet_outlined),
-            title: Text(s.loggerLevel),
-            subtitle: Text(loggerLevel.name),
-            trailing: SizedBox(
-              width: 200,
-              child: Slider.adaptive(
-                value: loggerLevelIndex.toDouble(),
-                onChanged: (value) async => ref
-                    .read(_loggerLevelProvider.notifier)
-                    .setIndex(value.toInt()),
-                max: loggerLevels.length - 1,
-                divisions: loggerLevels.length - 1,
-                label: loggerLevel.name,
-              ),
-            ),
-          ),
-        if (showDevTools)
-          ListTile(
-            leading: const Icon(Icons.network_check_outlined),
-            title: Text(s.networkTest),
-            onTap: () async => dio.getUri<dynamic>(testUrl),
-          )
+        if (showDevTools) const DevTools(),
       ],
     ),
   );
